requirements.txt
google-genai
embedchain
requests
beautifulsoup4
pillow
python-dotenv
chromadb
openai